requirements.txt
jsonschema>=4.17.0
pyyaml>=6.0
numpy>=1.21.0
pandas>=1.3.0
requests>=2.28.0
tqdm>=4.64.0